Apply file structure suggestions and code organizations from https://github.com/OpenSourceFellows/chatbot/pull/24#pullrequestreview-2918564801
