+++ src/controllers/index.js
@@ -1,2 +1,3 @@
 require('dotenv').config();
 module.exports.chatbotController = require('./chatbot.controller');
+module.exports.messagingController = require('./messaging.controller')
