+++ src/routes/messaging.routes.js
@@ -1,10 +1,10 @@
 const express = require('express')
+const messagingController = require('../controllers')
+
 const router = express.Router()
-const logger = require('../config/logger')
-const { handleIncomingMessage } = require('../services/messaging/incoming.messages.service')
+router.post('/webhook/incoming-message', messagingController.handleIncomingMessage)
 
-
-router.post('/webhook/incoming-message', handleIncomingMessage)
+module.exports = router;
 
 router.get('/', (req, res) => {
   res.send('📨 Messaging API is working');
