+++ src/utils/constants.js
@@ -1,0 +1,8 @@
+const SOURCES = {
+  TWILIO: 'twilio',
+  OPENEDX: 'openedx'
+}
+
+module.exports = {
+  SOURCES
+}
